Login/Signup Modals
Firebase.js changes: Updated the firebase.js file to export both the auth instance and the signOut method from firebase/auth. This change enables your other components to use these exports directly from firebase.js.

SignUp.js component: In the SignUp component, we introduced a new prop named handleClose. This prop is a function that gets called when a user has successfully signed up, closing the SignUp modal.

Login.js component: Similar changes were made in the Login component. We introduced a prop named handleClose, which is a function that is called when a user has successfully logged in, thereby closing the Login modal. There is a TODO comment left for future enhancement: "When sign up button is pressed make it close the Login modal completely". This indicates that you aim to also close the Login modal completely when the SignUp button is clicked, and this function will need to be implemented in the future.

Header.js component: The Header component was updated to handle the handleClose function passed as props to the SignUp and Login components. We also updated the way we import the signOut method from firebase.js and adjusted the implementation of the logout function.