User Context, State Management, User Blog Posts
In this commit, I've changed how I handle state management. Instead of managing state within components, I'm now using React's Context API. This makes the user's state available globally to all components.

A big part of this commit is the setup of a UserContext for user authentication state. This context uses Firebase's onAuthStateChanged function to listen for any changes to the user's login state and makes this information available to all components.

This meaning behind these changes is to set up the Create Post element of this project, which must check if the user is logged in to and if that user has the bool canPost == true before they can post. The post must also have a valid author (using their username).

TODO:

Fix the create post feature.
Improve the overall UI.
Look into refactoring the Login.js component to use UserContext.