Fixed routing issue
backend server.js and nginx route for the api was set to port 3000, Dockerfile.backend was set to port 8080.

fixed it by changing nginx and server.js to port 8080. also removed handling of the frontend from the backend.

some work on CreatePost, it is still not functional, need to sort out 400 post api error.